Clean up during PR review
